Split model and settings tests out to their own modules
